+++ karols-frontend/src/containers/ShopContainer.js
